[REFACTOR] winners 객체 삭제
