+++ src/main/java/racingcar/view/OutputView.java
@@ -2,7 +2,8 @@ package racingcar.view;
 
 import racingcar.domain.Car;
 import racingcar.domain.Cars;
-import racingcar.domain.Winners;
+
+import java.util.stream.Collectors;
 
 public class OutputView {
     public static final String DELIMTER = " : ";
@@ -24,7 +25,7 @@ public class OutputView {
     }
 
     private static void printCurrentPosition(Car car) {
-        StringBuilder currentPosition = new StringBuilder(car.toString());
+        StringBuilder currentPosition = new StringBuilder(car.getName());
         currentPosition.append(DELIMTER);
         for (int i = 0; i < car.getPosition(); i++) {
             currentPosition.append(POSITION_MARK);
@@ -39,8 +40,15 @@ public class OutputView {
         System.out.println();
     }
 
+    public static String getWinners() {
+        return Cars.winners.stream()
+                .map(Car::getName)
+                .collect(Collectors
+                        .joining(","));
+    }
+
     public static void printWinnerResult() {
-        System.out.println(Winners.getWinners() + "가 최종 우승했습니다.");
+        System.out.println(getWinners() + "가 최종 우승했습니다.");
     }
 
     public static void printErrorMessage(String errorMessage) {
