+++ src/test/java/racingcar/domain/FixedNumber.java
@@ -3,6 +3,7 @@ package racingcar.domain;
 import racingcar.domain.Generator.RandomGenerator;
 
 public class FixedNumber implements RandomGenerator {
+
     @Override
     public boolean isMovable() {
         return true;
